Refactored tests pass
