+++ .travis.yml
@@ -8,11 +8,10 @@ env:
   - secure: QWNGqrTcNUIdtpBKz1KMsx54SZ5OkQ1eK8SR4iJag4I3UXhT7MIwtbBrWyIEQvoC2in23gp5lPPiG842lPgKebOz+8GyJHYaE7SKPjPdpRfxPJSoDrJagArKNcsZrcZJkAr9jIpqV6Ap7p0TzzOWXsqYDbmLWxgrJOLHTLHi9QZc71vjP4wRhRBtz24xcUgSsRRUrM56FrxSy27zoem4AeJwRRVB52d1HwDbcHtbJ/3taVHFd4WLbdUIbwyOvXX3RGqVv2TxX9YoH+1S18mSoKy+OWulOQyrHc8pjPkz+cNyCYXE2Pu3m97e5c+NzFLsnfqYyKwROM49qGOvXbY0tziVYUAvgoJ1P+fsShKH5WKhv9J917R5QLf12iWKPI3/apBhXb5yGoFOerSxMUgJz4hSuKGTbg7nJ2hWJZVop7+c2LiSbtahghhmjJwlVzNSzTb6X6UayEgVlaqSF7wyq5pJmMXZ7fDxoooTyu4iDjRwYzxH1OrUWYz2Q9pHK7YAf9dHOgnrqaI6f+HQ9qt9auG9pGQ7Ep7tQVcikvWLZEgyqYLClagazqwoJImPrLrip3JVnb4lTudiwn2jKHYVgSjfEJn28aBf9ArtBpIvtnl5Rpe7fHrfeape5zUZlxC0lZ2sYWkoef/KooYWKk1iXOPA59RmswE6G8OZ+KmrV9s=
 install:
 - pip install -r requirements.txt
-- pip install Flask-Caching
 - pip install -e .
 - pip install codeclimate-test-reporter
 - pip install coverage pylint mock diff_cover coveralls
-script: coverage run --source=. -m unittest discover
+script: nosetests -v --with-coverage --cover-tests --cover-package=flask_mailgun
 after_success:
 - coveralls
 - codeclimate-test-reporter
