+++ flask_mailgun/__init__.py
@@ -4,12 +4,5 @@ Created on Wed Jan 27 21:48:14 2016
 
 @author: richard, yunxi
 """
-from flask import request, abort
-from flask_mailgun.api import MailGunAPI
-import os
-
-import tempfile
-from collections import defaultdict
-from decorator import decorator
-from functools import wraps
-
+from .api import MailGunAPI
+from .core import MailGun
